init_env move to ms_utils
