+++ examples/opensora_pku/examples/rec_video_vae.py
@@ -2,14 +2,12 @@ import argparse
 import logging
 import os
 import sys
-from typing import Tuple
 
 import numpy as np
 from tqdm import tqdm
 
 import mindspore as ms
 from mindspore import nn
-from mindspore.communication.management import get_group_size, get_rank, init
 
 mindone_lib_path = os.path.abspath("../../")
 sys.path.insert(0, mindone_lib_path)
@@ -17,113 +15,17 @@ sys.path.insert(0, mindone_lib_path)
 from mindone.utils.amp import auto_mixed_precision
 from mindone.utils.config import str2bool
 from mindone.utils.logger import set_logger
-from mindone.utils.seed import set_random_seed
 from mindone.visualize.videos import save_videos
 
 sys.path.append(".")
-from opensora.acceleration.parallel_states import (
-    get_sequence_parallel_state,
-    hccl_info,
-    initialize_sequence_parallel_state,
-)
+from opensora.acceleration.parallel_states import get_sequence_parallel_state, hccl_info
 from opensora.models.ae import getae_wrapper
 from opensora.models.ae.videobase.dataset_videobase import VideoDataset, create_dataloader
-
-# from opensora.models.ae.videobase.causal_vae.modeling_causalvae import TimeDownsample2x, TimeUpsample2x
 from opensora.models.ae.videobase.modules.updownsample import TrilinearInterpolate
+from opensora.utils.ms_utils import init_env
 from opensora.utils.utils import get_precision
 
 logger = logging.getLogger(__name__)
-
-
-def init_env(
-    mode: int = ms.GRAPH_MODE,
-    seed: int = 42,
-    distributed: bool = False,
-    max_device_memory: str = None,
-    device_target: str = "Ascend",
-    parallel_mode: str = "data",
-    precision_mode: str = None,
-    global_bf16: bool = False,
-    sp_size: int = 1,
-    jit_level: str = "O0",  # using kbk mode
-) -> Tuple[int, int, int]:
-    """
-    Initialize MindSpore environment.
-
-    Args:
-        mode: MindSpore execution mode. Default is 0 (ms.GRAPH_MODE).
-        seed: The seed value for reproducibility. Default is 42.
-        distributed: Whether to enable distributed training. Default is False.
-    Returns:
-        A tuple containing the device ID, rank ID and number of devices.
-    """
-    set_random_seed(seed)
-
-    if max_device_memory is not None:
-        ms.set_context(max_device_memory=max_device_memory)
-
-    if distributed:
-        ms.set_context(
-            mode=mode,
-            device_target=device_target,
-        )
-        if parallel_mode == "optim":
-            print("use optim parallel")
-            ms.set_auto_parallel_context(
-                parallel_mode=ms.ParallelMode.SEMI_AUTO_PARALLEL,
-                enable_parallel_optimizer=True,
-            )
-            init()
-            device_num = get_group_size()
-            rank_id = get_rank()
-        else:
-            init()
-            device_num = get_group_size()
-            rank_id = get_rank()
-            logger.debug(f"rank_id: {rank_id}, device_num: {device_num}")
-            ms.reset_auto_parallel_context()
-
-            ms.set_auto_parallel_context(
-                parallel_mode=ms.ParallelMode.DATA_PARALLEL,
-                gradients_mean=True,
-                device_num=device_num,
-            )
-
-        var_info = ["device_num", "rank_id", "device_num / 8", "rank_id / 8"]
-        var_value = [device_num, rank_id, int(device_num / 8), int(rank_id / 8)]
-        logger.info(dict(zip(var_info, var_value)))
-
-    else:
-        device_num = 1
-        rank_id = 0
-        ms.set_context(
-            mode=mode,
-            device_target=device_target,
-        )
-    if jit_level is not None:
-        if mode == 1:
-            print(f"Only graph mode supports jit_level! Will ignore jit_level {jit_level} in Pynative mode.")
-        else:
-            jit_dict = {"O0": "KBK", "O1": "DVM", "O2": "GE"}
-            print(f"Using jit_level: {jit_dict[jit_level]}")
-            ms.context.set_context(jit_config={"jit_level": jit_level})  # O0: KBK, O1:DVM, O2: GE
-    if global_bf16:
-        print("Using global bf16")
-        assert jit_level is not None and jit_level == "O2", "global_bf16 is supported in GE mode only!"
-        ms.set_context(
-            ascend_config={"precision_mode": "allow_mix_precision_bf16"}
-        )  # reset ascend precison mode globally
-
-    if precision_mode is not None and len(precision_mode) > 0:
-        ms.set_context(ascend_config={"precision_mode": precision_mode})
-
-    assert device_num >= sp_size and device_num % sp_size == 0, (
-        f"unable to use sequence parallelism, " f"device num: {device_num}, sp size: {sp_size}"
-    )
-    initialize_sequence_parallel_state(sp_size)
-
-    return rank_id, device_num
 
 
 def transform_to_rgb(x, rescale_to_uint8=True):
