+++ examples/opensora_pku/opensora/train/commons.py
@@ -1,126 +1,15 @@
 import argparse
 import logging
 import os
-from typing import Tuple
 
 import yaml
-from opensora.acceleration.parallel_states import get_sequence_parallel_state, initialize_sequence_parallel_state
 
-import mindspore as ms
 from mindspore import nn
-from mindspore.communication.management import get_group_size, get_rank, init
 from mindspore.nn.wrap.loss_scale import DynamicLossScaleUpdateCell
 
 from mindone.utils.config import str2bool
-from mindone.utils.seed import set_random_seed
 
 logger = logging.getLogger(__name__)
-
-
-def init_env(
-    mode: int = ms.GRAPH_MODE,
-    seed: int = 42,
-    distributed: bool = False,
-    max_device_memory: str = None,
-    device_target: str = "Ascend",
-    parallel_mode: str = "data",
-    mempool_block_size: str = "9GB",
-    global_bf16: bool = False,
-    strategy_ckpt_save_file: str = "",
-    optimizer_weight_shard_size: int = 8,
-    sp_size: int = 1,
-    jit_level: str = None,
-    enable_parallel_fusion: bool = False,
-) -> Tuple[int, int, int]:
-    """
-    Initialize MindSpore environment.
-
-    Args:
-        mode: MindSpore execution mode. Default is 0 (ms.GRAPH_MODE).
-        seed: The seed value for reproducibility. Default is 42.
-        distributed: Whether to enable distributed training. Default is False.
-    Returns:
-        A tuple containing the device ID, rank ID and number of devices.
-    """
-    set_random_seed(seed)
-    ms.set_context(mempool_block_size=mempool_block_size)
-
-    if max_device_memory is not None:
-        ms.set_context(max_device_memory=max_device_memory)
-    if enable_parallel_fusion:
-        ms.set_context(graph_kernel_flags="--enable_parallel_fusion --enable_expand_ops=AdamApplyOneWithDecayAssign")
-
-    if distributed:
-        ms.set_context(
-            mode=mode,
-            device_target=device_target,
-            ascend_config={"precision_mode": "allow_fp32_to_fp16"},  # ms2.2.23 parallel needs
-            # ascend_config={"precision_mode": "must_keep_origin_dtype"},  # TODO: tune
-        )
-        if parallel_mode == "optim":
-            print("use optim parallel")
-            ms.set_auto_parallel_context(
-                parallel_mode=ms.ParallelMode.SEMI_AUTO_PARALLEL,
-                parallel_optimizer_config={"optimizer_weight_shard_size": optimizer_weight_shard_size},
-                enable_parallel_optimizer=True,
-                strategy_ckpt_config={
-                    "save_file": strategy_ckpt_save_file,
-                    "only_trainable_params": False,
-                },
-            )
-            init()
-            device_num = get_group_size()
-            rank_id = get_rank()
-        elif parallel_mode == "data":
-            init()
-            device_num = get_group_size()
-            rank_id = get_rank()
-            logger.debug(f"rank_id: {rank_id}, device_num: {device_num}")
-            ms.reset_auto_parallel_context()
-
-            ms.set_auto_parallel_context(
-                parallel_mode=ms.ParallelMode.DATA_PARALLEL,
-                gradients_mean=True,
-                device_num=device_num,
-            )
-
-        var_info = ["device_num", "rank_id", "device_num / 8", "rank_id / 8"]
-        var_value = [device_num, rank_id, int(device_num / 8), int(rank_id / 8)]
-        logger.info(dict(zip(var_info, var_value)))
-
-    else:
-        device_num = 1
-        rank_id = 0
-        ms.set_context(
-            mode=mode,
-            device_target=device_target,
-            ascend_config={"precision_mode": "allow_fp32_to_fp16"},  # TODO: tune for better precision
-        )
-
-    if jit_level is not None:
-        if mode == 1:
-            print(f"Only graph mode supports jit_level! Will ignore jit_level {jit_level} in Pynative mode.")
-        else:
-            jit_dict = {"O0": "KBK", "O1": "DVM", "O2": "GE"}
-            print(f"Using jit_level: {jit_dict[jit_level]}")
-            ms.context.set_context(jit_config={"jit_level": jit_level})  # O0: KBK, O1:DVM, O2: GE
-    if global_bf16:
-        print("Using global bf16")
-        assert jit_level is not None and jit_level == "O2", "global_bf16 is supported in GE mode only!"
-        ms.set_context(
-            ascend_config={"precision_mode": "allow_mix_precision_bf16"}
-        )  # reset ascend precison mode globally
-
-    assert device_num >= sp_size and device_num % sp_size == 0, (
-        f"unable to use sequence parallelism, " f"device num: {device_num}, sp size: {sp_size}"
-    )
-    initialize_sequence_parallel_state(sp_size)
-    if get_sequence_parallel_state():
-        assert (
-            parallel_mode == "data"
-        ), f"only support seq parallelism with parallel mode `data`, but got `{parallel_mode}`"
-
-    return rank_id, device_num
 
 
 def _check_cfgs_in_parser(cfgs: dict, parser: argparse.ArgumentParser):
